requirements.txt
fastapi
uvicorn
faiss-cpu
aiohttp
torch
transformers
datasets
accelerate
bitsandbytes
lfs
pillow
peft
soundfile
qwen_tts